Make Sha256Hash public, with private impl
