+++ include/ccf/tx.h
@@ -3,8 +3,8 @@
 #pragma once
 
 #include "ccf/ccf_assert.h"
+#include "ccf/crypto/sha256_hash.h"
 #include "ccf/tx_id.h"
-#include "ccf/crypto.h"
 #include "kv/kv_types.h"
 #include "kv/untyped_map.h"
