+++ src/apps/js_v8/tmpl/ccf_global.cpp
@@ -4,7 +4,7 @@
 
 #include "ccf/ds/logger.h"
 #include "consensus.h"
-#include "crypto.h"
+#include "apps/js_v8/tmpl/crypto.h"
 #include "crypto/entropy.h"
 #include "crypto/key_wrap.h"
 #include "crypto/rsa_key_pair.h"
