+++ src/node/history.h
@@ -199,7 +199,7 @@ namespace ccf
     {
       return {
         {term_of_last_version, version},
-        crypto::Sha256Hash(CBuffer(std::to_string(version))),
+        crypto::Sha256Hash(std::to_string(version)),
         term_of_next_version};
     }
 
@@ -802,7 +802,7 @@ namespace ccf
 
     void append(const std::vector<uint8_t>& data) override
     {
-      crypto::Sha256Hash rh({data.data(), data.size()});
+      crypto::Sha256Hash rh(data);
       log_hash(rh, APPEND);
       std::lock_guard<std::mutex> guard(state_lock);
       replicated_state_tree.append(rh);
